refactor features component
